pyproject.toml
[project]
name = "duvidnn"
version = "0.0.3"
authors = [
  { name="Eachan Johnson", email="eachan.johnson@crick.ac.uk" },
]
description = "Deep learning with uncertainty for general purpose, chemistry, and taxonomic tasks."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = [
  "ai",
  "active-learning", 
  "bayesian-optimization",
  "data",
  "deep-learning",
  "machine-learning",
]

classifiers = [  
  
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "carabiner-tools[pd,mpl]>=0.0.4",
  "datasets>=2.15.0",
  "duvida>=0.0.3",
  "huggingface_hub",
  "numpy",
  "torch>=2.4",
  "torchvision",
  "lightning",
  "tensorboard",
  "tqdm",
]

[project.optional-dependencies]
transformers = ["torch>=2.4", "torchvision", "transformers"]
chem = ["chemprop>=2.1.0", "schemist>=0.0.4.post1"]
splits = ["umap-learn", "scipy", "schemist>=0.0.4.post1", "tdigest>=0.5", "faiss-cpu>=1.11"]
splits_gpu = ["umap-learn", "scipy", "schemist>=0.0.4.post1", "tdigest>=0.5", "faiss-gpu"]

[project.urls]
"Homepage" = "https://github.com/scbirlab/duvidnn"
"Bug Tracker" = "https://github.com/scbirlab/duvidnn/issues"

[project.scripts] 
duvida = "duvida.cli:main"

[tool.setuptools]
#packages = ["duvida"]
# If there are data files included in your packages that need to be
# installed, specify them here.
package-data = {duvida = ["data/*.json"]}

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[test]
deps = ["pytest>=7.4", "pytest-cov", "duvida[chem,jax,torch,transformers,splits]"] 
